+++ backend/src/main/java/ca/bc/gov/app/security/ApiAuthorizationCustomizer.java
@@ -85,17 +85,6 @@ public class ApiAuthorizationCustomizer implements Customizer<AuthorizeExchangeS
         .hasAnyRole(ApplicationConstant.ROLE_EDITOR,
             ApplicationConstant.ROLE_ADMIN);
 
-    // Viewer, editor, admin, BCeIDBusiness, BCSC and service users can GET from the countries
-    // endpoint
-    authorize
-        .pathMatchers(HttpMethod.GET, "/api/countries/**")
-        .hasAnyRole(ApplicationConstant.ROLE_VIEWER,
-            ApplicationConstant.ROLE_EDITOR,
-            ApplicationConstant.ROLE_ADMIN,
-            ApplicationConstant.USERTYPE_BCEIDBUSINESS_USER,
-            ApplicationConstant.USERTYPE_BCSC_USER,
-            ApplicationConstant.USERTYPE_SERVICE_USER);
-
     // Only editor and admin can POST to the clients submissions endpoint with a specific id
     authorize
       .pathMatchers(HttpMethod.POST, "/api/clients/submissions/{id:[0-9]+}")
