+++ backend/src/main/java/ca/bc/gov/app/validator/client/ClientSubmitRequestValidator.java
@@ -14,8 +14,6 @@ import ca.bc.gov.app.repository.client.ClientTypeCodeRepository;
 import ca.bc.gov.app.repository.client.DistrictCodeRepository;
 import ca.bc.gov.app.service.client.ClientService;
 import java.time.LocalDate;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
